`sizeof(::Array)`: typeassert `::Int` to help abstract type inference
Fix a regression where abstract return type inference of
`sizeof(::Array)` used to produce `Any` after Julia v1.11.

+++ base/array.jl
@@ -232,7 +232,7 @@ function elsize(::Type{Ptr{T}}) where T
 end
 elsize(::Type{Union{}}, slurp...) = 0
 
-sizeof(a::Array) = length(a) * elsize(typeof(a)) # n.b. this ignores bitsunion bytes, as a historical fact
+sizeof(a::Array) = length(a)::Int * elsize(typeof(a)) # n.b. this ignores bitsunion bytes, as a historical fact
 
 function isassigned(a::Array, i::Int...)
     @inline
